add autocomplete for school zones
