+++ package.json
@@ -14,6 +14,7 @@
     "js-cookie": "^2.2.0",
     "react": "^16.5.2",
     "react-accessible-accordion": "^2.4.4",
+    "react-autocomplete": "^1.8.1",
     "react-dom": "^16.5.2",
     "react-emotion": "^9.2.10",
     "react-load-script": "0.0.6",
