--- src/constants/newWorkOrderFormConfig.js
@@ -1,233 +1,0 @@
-export const ASSET_TYPE_OPTIONS = [
-  "Signal",
-  "Camera",
-  "School Beacon",
-  "Hazard Flasher",
-  "Digital Messaging Sign (DMS)",
-  "Sensor",
-  "Other / No Asset"
-];
-
-export const SCHOOL_ZONE_OPTIONS = [
-  { id: "", name: "Select" },
-  { id: "5b63411dbe8a97076888173f", name: "ALLAN" },
-  { id: "5b63411cbe8a970768881605", name: "ALLAN COLLEGE PREP, IDEA SCHOOL" },
-  { id: "5b63411dbe8a9707688816b1", name: "ALLISON" },
-  { id: "5b63411cbe8a970768881665", name: "ANDERSON MILL, GRISHAM" },
-  { id: "5b63411dbe8a9707688816eb", name: "ANDREWS" },
-  {
-    id: "5b63411cbe8a97076888162f",
-    name: "ANN RICHARDS SCH FOR YOUNG WOMEN LEADERS"
-  },
-  { id: "5b63411bbe8a9707688815ff", name: "ANN RICHARDS SCHOOL FOR GIRLS" },
-  { id: "5b63411dbe8a970768881715", name: "AUSTIN" },
-  { id: "5b63411dbe8a9707688816c3", name: "AUSTIN ACHEIVE" },
-  { id: "5b63411bbe8a9707688815fb", name: "AUSTIN DISCOVERY SCHOOL FM 969" },
-  { id: "5b63411cbe8a97076888161f", name: "BAILEY" },
-  { id: "5b63411cbe8a970768881621", name: "BALDWIN" },
-  { id: "5b63411cbe8a97076888161d", name: "BANNOCKBURN CHRISTIAN ACADEMY" },
-  { id: "5b63411cbe8a97076888162b", name: "BARANOFF" },
-  { id: "5b63411dbe8a97076888170b", name: "BARRINGTON" },
-  { id: "5b63411cbe8a970768881607", name: "BARRON" },
-  { id: "5b63411cbe8a970768881659", name: "BARTON HILLS" },
-  { id: "5b63411dbe8a9707688816c1", name: "BECKER" },
-  { id: "5b63411cbe8a97076888163b", name: "BEDICHEK" },
-  { id: "5b63411dbe8a9707688816e5", name: "BERTHA SADLER, PECAN SPRINGS" },
-  { id: "5b63411dbe8a970768881727", name: "BLACKSHEAR" },
-  { id: "5b63411dbe8a9707688816e1", name: "BLANTON" },
-  { id: "5b63411dbe8a9707688816d9", name: "BLUEBONNET TRAIL" },
-  { id: "5b63411bbe8a970768881603", name: "BLUFF SPRINGS IDEA SCHOOL" },
-  { id: "5b63411cbe8a970768881647", name: "BOONE" },
-  { id: "5b63411cbe8a97076888161b", name: "BOWIE" },
-  { id: "5b63411cbe8a970768881675", name: "BRENTWOOD" },
-  { id: "5b63411dbe8a9707688816d1", name: "BRENTWOOD OAKS" },
-  { id: "5b63411cbe8a970768881651", name: "BRIDGE POINT" },
-  { id: "5b63411dbe8a970768881733", name: "BROOKE" },
-  { id: "5b63411dbe8a9707688816f5", name: "BROWN" },
-  { id: "5b63411dbe8a97076888171f", name: "BRYKER WOODS" },
-  { id: "5b63411cbe8a97076888167f", name: "BURNET" },
-  { id: "5b63411cbe8a97076888169b", name: "BUSH/CANYON RIDGE" },
-  { id: "5b63411dbe8a970768881723", name: "CAMPBELL" },
-  { id: "5b63411cbe8a97076888168d", name: "CANYON CREEK" },
-  { id: "5b63411cbe8a97076888168b", name: "CANYON VISTA" },
-  { id: "5b63411cbe8a970768881685", name: "CARAWAY" },
-  { id: "5b63411cbe8a970768881625", name: "CASEY" },
-  { id: "5b63411dbe8a97076888171d", name: "CASIS" },
-  { id: "5b63411cbe8a970768881655", name: "CEDAR CREEK" },
-  { id: "5b63411cbe8a970768881643", name: "CLAYTON" },
-  { id: "5b63411dbe8a9707688816fb", name: "COOK" },
-  { id: "5b63411dbe8a970768881701", name: "COPPERFIELD" },
-  { id: "5b63411cbe8a970768881649", name: "COVINGTON" },
-  { id: "5b63411cbe8a970768881627", name: "COWAN" },
-  { id: "5b63411cbe8a970768881635", name: "CROCKETT" },
-  { id: "5b63411cbe8a970768881633", name: "CUNNINGHAM, EDEN PARK ACADEMY" },
-  { id: "5b63411cbe8a970768881691", name: "DAVIS" },
-  { id: "5b63411cbe8a9707688816a5", name: "DAWSON" },
-  { id: "5b63411cbe8a9707688816a1", name: "DEL VALLE" },
-  { id: "5b63411dbe8a9707688816d3", name: "DELCO, DESSAU" },
-  { id: "5b63411dbe8a9707688816cd", name: "DESSAU" },
-  { id: "5b63411cbe8a970768881613", name: "DITTMAR REC CENTER" },
-  { id: "5b63411dbe8a9707688816f7", name: "DOBIE" },
-  { id: "5b63411cbe8a970768881679", name: "DOSS, MURCHISON" },
-  {
-    id: "5b63411dbe8a970768881713",
-    name: "EASTSIDE MEMORIAL AT THE JOHNSTON CAMPUS"
-  },
-  { id: "5b63411dbe8a9707688816b9", name: "FULMORE" },
-  { id: "5b63411dbe8a9707688816a7", name: "GALINDO" },
-  { id: "5b63411dbe8a9707688816c9", name: "GARCIA" },
-  { id: "5b63411dbe8a970768881711", name: "GARZA INDEPENDENCE" },
-  { id: "5b63411cbe8a970768881623", name: "GORZYCKI" },
-  { id: "5b63411dbe8a970768881731", name: "GOVALLE" },
-  { id: "5b63411dbe8a9707688816ff", name: "GRAHAM" },
-  { id: "5b63411cbe8a97076888166b", name: "GRANDVIEW HILLS" },
-  { id: "5b63411dbe8a9707688816c5", name: "GUERRO-THOMPSON" },
-  { id: "5b63411cbe8a970768881677", name: "GULLETT" },
-  { id: "5b63411cbe8a97076888169f", name: "HARMONY SCHOOL OF EXCELLENCE" },
-  { id: "5b63411dbe8a9707688816df", name: "HARRIS" },
-  { id: "5b63411dbe8a9707688816f1", name: "HART" },
-  { id: "5b63411cbe8a970768881671", name: "HIGHLAND PARK" },
-  { id: "5b63411cbe8a970768881657", name: "HILL COUNTRY" },
-  { id: "5b63411cbe8a970768881683", name: "HILL, ANDERSON" },
-  { id: "5b63411dbe8a9707688816af", name: "HOUSTON" },
-  { id: "5b63411dbe8a97076888170f", name: "HYDE PARK BAPTIST" },
-  { id: "5b63411bbe8a9707688815f7", name: "JOE LEE JOHNSON" },
-  { id: "5b63411dbe8a9707688816cf", name: "JOHN B CONNALLY" },
-  { id: "5b63411cbe8a970768881687", name: "JOLLYVILLE" },
-  { id: "5b63411dbe8a9707688816e3", name: "JORDAN" },
-  { id: "5b63411cbe8a970768881631", name: "JOSLIN" },
-  { id: "5b63411dbe8a970768881729", name: "KEALING" },
-  { id: "5b63411cbe8a970768881641", name: "KIKER" },
-  { id: "5b63411bbe8a9707688815f9", name: "KIPP AUSTIN CONNECTIONS" },
-  { id: "5b63411cbe8a97076888163f", name: "KOCUREK" },
-  { id: "5b63411dbe8a9707688816ed", name: "L.B.J" },
-  { id: "5b63411dbe8a9707688816ab", name: "LANGFORD" },
-  { id: "5b63411dbe8a9707688816f9", name: "LANIER" },
-  { id: "5b63411cbe8a970768881689", name: "LAUREL MOUNTAIN" },
-  { id: "5b63411dbe8a970768881721", name: "LEE" },
-  { id: "5b63411dbe8a9707688816b3", name: "LINDER" },
-  { id: "5b63411cbe8a970768881661", name: "LIVE OAK EL/ DEER PARK MS" },
-  { id: "5b63411dbe8a970768881709", name: "MAPLEWOOD" },
-  { id: "5b63411dbe8a97076888173b", name: "MARTIN" },
-  { id: "5b63411dbe8a970768881719", name: "MATHEWS" },
-  { id: "5b63411dbe8a9707688816d7", name: "MCBEE" },
-  { id: "5b63411cbe8a970768881673", name: "MCCALLUM" },
-  { id: "5b63411cbe8a97076888165d", name: "MCNEIL" },
-  { id: "5b63411dbe8a970768881739", name: "METZ" },
-  { id: "5b63411cbe8a970768881645", name: "MILLS" },
-  { id: "5b63411dbe8a97076888172d", name: "NORMAN" },
-  { id: "5b63411cbe8a970768881697", name: "NORTHWEST" },
-  { id: "5b63411dbe8a97076888171b", name: "O HENRY" },
-  { id: "5b63411cbe8a97076888164d", name: "OAK HILL" },
-  { id: "5b63411dbe8a97076888172b", name: "OAK SPRINGS" },
-  { id: "5b63411cbe8a970768881639", name: "ODOM" },
-  { id: "5b63411dbe8a970768881737", name: "ORTEGA" },
-  { id: "5b63411cbe8a970768881609", name: "PADRON" },
-  { id: "5b63411dbe8a9707688816ad", name: "PALM" },
-  { id: "5b63411cbe8a97076888166d", name: "PATSY SOMMER" },
-  { id: "5b63411cbe8a97076888164b", name: "PATTON" },
-  { id: "5b63411dbe8a9707688816e9", name: "PEARCE" },
-  { id: "5b63411dbe8a9707688816e7", name: "PEARCE, ANDREWS" },
-  { id: "5b63411dbe8a970768881717", name: "PEASE" },
-  { id: "5b63411cbe8a97076888160d", name: "PECAN SPRINGS" },
-  { id: "5b63411dbe8a9707688816d5", name: "PICKLE" },
-  { id: "5b63411cbe8a970768881681", name: "PILLOW" },
-  { id: "5b63411dbe8a9707688816c7", name: "PIONEER CROSSING" },
-  { id: "5b63411dbe8a9707688816a9", name: "PLEASANT HILL" },
-  { id: "5b63411cbe8a97076888165f", name: "POND SPRINGS" },
-  { id: "5b63411dbe8a970768881703", name: "PREMIER AT LANIER" },
-  { id: "5b63411cbe8a970768881669", name: "PURPLE SAGE" },
-  { id: "5b63411dbe8a9707688816db", name: "REAGAN" },
-  { id: "5b63411cbe8a97076888160b", name: "REAGAN" },
-  { id: "5b63411cbe8a970768881619", name: "REGENTS SCHOOL OF AUSTIN" },
-  { id: "5b63411dbe8a9707688816ef", name: "REILLY" },
-  { id: "5b63411dbe8a9707688816dd", name: "RIDGETOP" },
-  { id: "5b63411cbe8a97076888166f", name: "RIVER PLACE" },
-  { id: "5b63411cbe8a970768881663", name: "RIVER RIDGE" },
-  { id: "5b63411cbe8a9707688816a3", name: "RODRIGUEZ" },
-  { id: "5b63411cbe8a970768881617", name: "RUTLEDGE" },
-  { id: "5b63411dbe8a97076888173d", name: "SANCHEZ" },
-  { id: "5b63411dbe8a97076888172f", name: "SIMS" },
-  { id: "5b63411cbe8a970768881629", name: "SMALL" },
-  { id: "5b63411cbe8a97076888168f", name: "SPICEWOOD" },
-  { id: "5b63411cbe8a970768881615", name: "ST AUSTIN CATHOLIC SCHOOL" },
-  { id: "5b63411dbe8a9707688816bd", name: "ST ELMO" },
-  { id: "5b63411dbe8a9707688816cb", name: "ST FRANCIS" },
-  { id: "5b63411dbe8a9707688816b7", name: "ST IGNATIUS" },
-  { id: "5b63411cbe8a97076888167b", name: "ST LOUIS SCHOOL" },
-  { id: "5b63411cbe8a97076888164f", name: "ST MICHAELS" },
-  { id: "5b63411dbe8a970768881725", name: "ST. MARTIN LUTHERAN SCHOOL" },
-  { id: "5b63411cbe8a97076888169d", name: "STEINER RANCH" },
-  { id: "5b63411cbe8a970768881637", name: "STRICKLAND CHRISTIAN SCHOOL" },
-  { id: "5b63411cbe8a970768881693", name: "SUMMIT" },
-  { id: "5b63411cbe8a97076888162d", name: "SUNSET VALLEY" },
-  { id: "5b63411dbe8a9707688816b5", name: "TRAVIS" },
-  { id: "5b63411dbe8a9707688816bb", name: "TRAVIS HEIGHTS" },
-  { id: "5b63411dbe8a97076888170d", name: "UT ELEMENTARY CHARTER SCHOOL" },
-  { id: "5b63411dbe8a9707688816fd", name: "WALNUT CREEK" },
-  { id: "5b63411dbe8a9707688816f3", name: "WEBB" },
-  { id: "5b63411cbe8a97076888160f", name: "WEBB, BROWN" },
-  { id: "5b63411cbe8a970768881699", name: "WELLS BRANCH" },
-  { id: "5b63411bbe8a9707688815fd", name: "WEST RIDGE" },
-  { id: "5b63411cbe8a970768881653", name: "WESTLAKE" },
-  { id: "5b63411cbe8a970768881695", name: "WESTVIEW, PARMER" },
-  { id: "5b63411cbe8a970768881667", name: "WESTWOOD" },
-  { id: "5b63411dbe8a9707688816bf", name: "WIDEN" },
-  { id: "5b63411cbe8a97076888163d", name: "WILLIAMS" },
-  { id: "5b63411cbe8a970768881611", name: "WINDERMERE" },
-  { id: "5b63411dbe8a970768881707", name: "WINN" },
-  { id: "5b63411dbe8a970768881705", name: "WOOLDRIDGE" },
-  { id: "5b63411bbe8a970768881601", name: "WOOTEN, BURNET" },
-  { id: "5b63411cbe8a97076888167d", name: "WOOTEN, REEDEMER" },
-  { id: "5b63411dbe8a970768881735", name: "ZAVALA" },
-  { id: "5b63411cbe8a97076888165b", name: "ZILKER" }
-];
-
-export const WORK_TYPE_TROUBLE_CALL_OPTIONS = [
-  "LED Out",
-  "Communication Failure",
-  "Detection Failure",
-  "Digtess",
-  "Knockdown",
-  "Push Button Not Working",
-  "Signal Out or on Flash",
-  "Timing Issue",
-  "Visibility Issue",
-  "Other"
-];
-
-export const WORK_TYPE_SCHEDULED_WORK_OPTIONS = [
-  "Call-Back (Test Monitors and Cabinets)",
-  "On-Call Person (OCP)",
-  "Detection - Loop Tie-Ins and Research",
-  "Detection - Other",
-  "Installation - Install Heads",
-  "Installation - APS",
-  "Installation - Build Signal Heads",
-  "Installation - Cabinet",
-  "Installation - Cable",
-  "Installation - Camera",
-  "Installation - Detection - Video",
-  "Installation - Fiber",
-  "Installation - Louvers",
-  "Installation - Other",
-  "Installation - Ped Countdown Inserts",
-  "Installation - Poles and/or Arms",
-  "Installation - Sensor",
-  "Preventative Maintenance",
-  "Record APS Audio",
-  "Misc - School Zone Transfer",
-  "Misc - Test Cabinets",
-  "Misc - Traffic Count",
-  "Misc - Training",
-  "Misc - Turn On Signal",
-  "Misc - Assist TxDOT",
-  "Pole Removal",
-  "Repair - Fiber",
-  "Repair - CCTV Camera",
-  "Repair - Detection - Loop",
-  "Repair - Detection - Video",
-  "Repair - Preemption",
-  "Repair - UPS Issues (Batteries)",
-  "Knockdown Follow-Up"
-];
